+++ src/Avalonia.Controls.WebView/NativeWebView.cs
@@ -317,9 +317,7 @@ namespace Avalonia.Xpf.Controls
             Task.FromException<Stream>(new PlatformNotSupportedException());
 
         /// <inheritdoc cref="PrintToPdfStreamAsync()"/>
-        [UnsupportedOSPlatform("macos")]
-        [UnsupportedOSPlatform("ios")]
-        public Task<Stream> PrintToPdfStreamAsync(AvPlatform.WebViewPrintSettings printSettings) => TryGetAdapter() is Core.IWebViewWithPrintWithOptions adapter ?
+        public Task<Stream> PrintToPdfStreamAsync(AvPlatform.WebViewPrintSettings printSettings) => TryGetAdapter() is Core.IWebViewWithPrint adapter ?
             adapter.PrintToPdfStreamAsync(printSettings) :
             Task.FromException<Stream>(new PlatformNotSupportedException());